Flush pending cq events when processing DISCONNECTED cm event
Upon reception of CM RDMA_CM_EVENT_DISCONNECTED event (eg. when remote
side closes the connection), msk_cma_event_handler() disables the
polling of completion events for the transport (remove completion
channel fd from the cq's epoll_event) but it doesn't flush the last
pending work completions that may still be in the completion queue.

In some cases, under heavy load on the server, we saw the disconnect
cm event initiated by the client has overtaken the last send completion
on the server side (completion for the last reply sent, the one that
triggered the disconnection from client).

Not receiving the last send completion has impact on the release of
resources on the server and replenishment of the shared receive queue.

This commit adds a call to msk_cq_event_handler() in the
RDMA_CM_EVENT_DISCONNECTED case (similarly to what is done in
msk_flush_buffers()), prior to changing the transport state
and calling the disconnect callback.